refactor: update installation scripts and add user prompts
- Replaced the uninstall script call in install.sh with a local-install.sh source.
- Added a user prompt in install.sh to confirm deletion of installation files after setup.
- Enhanced kde.sh to define additional environment paths for templates and documentation.
- Updated local-install.sh to copy documentation and templates to the installation directory.
- Introduced a new .gitignore file in the templates/init directory to manage ignored files effectively.

+++ install.sh
@@ -4,14 +4,16 @@ git clone https://github.com/r82wei/KDE-cli.git
 
 cd KDE-cli
 
-source ./uninstall.sh
-
 # 安裝 kde 腳本
-mkdir -p /usr/local/lib/kde
-cp -r kde.sh /usr/local/lib/kde/
-cp -r ./scripts /usr/local/lib/kde/
-ln -s /usr/local/lib/kde/kde.sh /usr/local/bin/kde
-
+source ./local-install.sh
 
 cd ..
-rm -rf KDE-cli
+# 詢問使用者是否要刪除安裝檔案
+read -p "是否要刪除安裝檔案？(y/n): " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+    rm -rf KDE-cli
+    echo "安裝檔案已刪除"
+else
+    echo "保留安裝檔案"
+fi
